// src/app/models/user.ts
export class User {
    firstname: String;
    lastname: String;
    mobilenumber: String;
    email: String;
    dob: String;
    gender: String;
    address: Object;
    userRole: Number;
    userLog: Number;
    userActive: Boolean;
    password: String;
    last_login: String;
    created_at: Date;
    maritalstatus: String;
}